OTH: Splitting classes and adding absolute paths
