--- Interface/react/src/classes/size.ts
@@ -1,10 +1,0 @@
-/*
-
-This program has been developed by students from the bachelor Computer Science at
-Utrecht University within the Software Project course.
-© Copyright Utrecht University (Department of Information and Computing Sciences)
-
- */
-
-/** The size of the overlay, used to scale the drawing of the bounding boxes. */
-export type size = { left: number; top: number; width: number; height: number }
